Switch to advanced ops to avoid a breaking change for now
added a note that we do still want to rename next time we make a breaking
change (potentially removing owner for ex)
