+++ src/document.rs
@@ -19,7 +19,6 @@ use futures::Future;
 use itertools::{Either, EitherOrBoth, Itertools};
 
 pub mod advanced;
-pub mod unmanaged;
 
 /// List of users and groups that should have access to decrypt a document.
 #[derive(Clone, Debug, Eq, Hash, PartialEq)]
